Split events into multiple files
