+++ events/models.ts
@@ -1,0 +1,11 @@
+/**
+ * All possible colors of an event.
+ */
+export type EventColor =
+  | "#EF6351"
+  | "#CB9CF2"
+  | "#88BDEA"
+  | "#72B01D"
+  | "#F7B2BD"
+  | "#F4845F"
+  | "#F6BD60";
